% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_api.R
\name{create_hdf5_file}
\alias{create_hdf5_file}
\title{Create an HDF5 File}
\usage{
create_hdf5_file(hdf5_file)
}
\arguments{
\item{hdf5_file}{A character string specifying the path and name of the HDF5
file to be created.}
}
\value{
No return value; the function creates the HDF5 file at the specified
location.
}
\description{
This function creates a new, empty HDF5 file at the specified location.
}
\examples{
# Create an empty HDF5 file
hdf5_file <- tempfile()
create_hdf5_file(hdf5_file)


}
